Finishing touches on this project - will abandon this repo soon

+++ src/instruments/guitar/MusicNote.java
@@ -163,7 +163,7 @@ public class MusicNote {
 		return duration;
 	}
 
-	public double getVelocity() {
+	public int getVelocity() {
 		return velocity;
 	}
